+++ features/support/env.rb
@@ -3,12 +3,13 @@
 $api_key = 'TEST_API_KEY'
 
 # Set which test fixture should be used
-ENV['APP_FIXTURE_DIR'] ||= 'features/fixtures/app'
-ENV['NDK_FIXTURE_DIR'] ||= 'features/fixtures/ndkapp'
-ENV['RN_FIXTURE_DIR'] ||= 'features/fixtures/rn063/android'
-ENV['RN_MONOREPO_FIXTURE_DIR'] ||= 'features/fixtures/rn-monorepo/abc/android'
-ENV['UNITY_2018_FIXTURE_DIR'] ||= 'features/fixtures/unity_2018/example'
-ENV['UNITY_2019_FIXTURE_DIR'] ||= 'features/fixtures/unity_2019'
+ENV["APP_FIXTURE_DIR"] ||= "features/fixtures/app"
+ENV["NDK_FIXTURE_DIR"] ||= "features/fixtures/ndkapp"
+ENV["RN_FIXTURE_DIR"] ||= "features/fixtures/rn063/android"
+ENV["LIB_FIXTURE_DIR"] ||= "features/fixtures/mylib"
+ENV["RN_MONOREPO_FIXTURE_DIR"] ||= "features/fixtures/rn-monorepo/abc/android"
+ENV["UNITY_2018_FIXTURE_DIR"] ||= "features/fixtures/unity_2018/example"
+ENV["UNITY_2019_FIXTURE_DIR"] ||= "features/fixtures/unity_2019"
 
 # set defaults for versions
 ENV['AGP_VERSION'] ||= '4.1.0' # default to latest