+++ CHANGELOG.md
@@ -3,14 +3,14 @@
 * Address task dependency warning when using APK splits
   [#408](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/408)
 
-* Address task dependency warning when using APK splits
-  [#412](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/412)
-
 * Custom Dexguard paths are resolved correctly (relative to the project)
   [#420](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/420)
 
 * Release-metadata tasks are always available for configured projects
   [#422](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/422)
+
+* Support extracting SO files from bugsnag AARs for library modules
+  [#423](https://github.com/bugsnag/bugsnag-android-gradle-plugin/pull/423)
 
 ## 5.7.8 (2021-07-22)